requirements.txt
Flask
python-dotenv
google-generativeai
gTTS
pdfplumber
Werkzeug